requirements.txt
pandas==2.2.1
numpy==1.26.4
thefuzz[speedup]==0.22.1
python-dateutil==2.9.0.post0
pydantic==2.6.4
openai==1.14.2
python-dotenv==1.0.1
httpx==0.25.2
Flask==3.0.3
fuzzywuzzy
python-Levenshtein
matplotlib 